# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "(Quiz) Server API - Swagger",
    "description": "Account Admin: admin - 12345 <br /> <br /> Account User: user - 12345",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "definitions": {
...
    "QuizDTO": {
      "type": "object",
      "required": ["code"],
      "properties": {
        "code": {
          "type": "string",
          "example": "MAS291"
        },
        "name": {
          "type": "string",
          "example": "Statistics and Probability"
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/quiz/get-all": {
      "get": {
        "tags": ["Quiz Controller"],
        "summary": "Get the list of quizs",
        "operationId": "getAllQuiz",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully - List of quíz",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/quiz/get-one/{id}": {
      "get": {
        "tags": ["Quiz Controller"],
        "summary": "Get quiz by id",
        "produces": ["application/xml", "application/json"],
        "operationId": "getQuizById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of quiz to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully - Quiz Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID! ID Must be a number"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Quiz not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/quiz/create": {
      "post": {
        "tags": ["Quiz Controller"],
        "summary": "Create quiz",
        "operationId": "createQuiz",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "QuizDTO",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              }
            }
          },
          "400": {
            "description": "Quiz code already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/quiz/update": {
      "put": {
        "tags": ["Quiz Controller"],
        "summary": "Update quiz",
        "operationId": "updateQuiz",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "QuizDTO",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/definitions/QuizDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              }
            }
          },
          "400": {
            "description": "Quiz code already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Quiz not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/quiz/delete/{id}": {
      "delete": {
        "tags": ["Quiz Controller"],
        "summary": "Delete quiz",
        "produces": ["application/xml", "application/json"],
        "operationId": "deleteQuizById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of quiz to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuizDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Quiz not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}